requirements.txt
# Core runtime deps
fastapi==0.121.0
uvicorn==0.38.0
SQLAlchemy==2.0.44
asyncpg==0.30.0
pydantic==2.12.3
python-dotenv==1.2.1
prometheus_client==0.23.1
APScheduler==3.11.1

# gRPC / protobuf
grpcio==1.76.0
grpcio-tools==1.76.0  # needed if you generate protos during build; optional at runtime
protobuf==6.33.0

# logging / utils
structlog==25.5.0
colorlog==6.10.1
colorama==0.4.6

# misc
anyio==4.11.0
h11==0.16.0
idna==3.11
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
pydantic>=2.0
slowapi==0.1.9
pydantic-settings==2.12.0

